+++ tests/integration/tests/plugins/zod.test.ts
@@ -458,6 +458,47 @@ describe('Zod plugin tests', () => {
         expect(schema.safeParse({ arr: [1, 2, 3] }).success).toBeTruthy();
     });
 
+    it('refinement with path', async () => {
+        const model = `
+        datasource db {
+            provider = 'postgresql'
+            url = env('DATABASE_URL')
+        }
+        
+        generator js {
+            provider = 'prisma-client-js'
+        }
+
+        plugin zod {
+            provider = "@core/zod"
+        }
+
+        model M {
+            id Int @id @default(autoincrement())
+            arr Int[]
+
+            @@validate(!isEmpty(arr), 'condition1', ['array'])
+            @@validate(has(arr, 1), 'condition2', ['arr'])
+            @@validate(hasEvery(arr, [1, 2]), 'condition3', ['arr', 'every'])
+            @@validate(hasSome(arr, [1, 2]), 'condition4', ['arr', 'some'])
+        }
+        `;
+
+        const { zodSchemas } = await loadSchema(model, { addPrelude: false, pushDb: false });
+
+        const schema = zodSchemas.models.MCreateSchema;
+        expect(schema.safeParse({}).error.issues[0].path).toEqual(['array']);
+        expect(schema.safeParse({ arr: [] }).error.issues[0].path).toEqual(['array']);
+        expect(schema.safeParse({ arr: [3] }).error.issues[0].path).toEqual(['arr']);
+        expect(schema.safeParse({ arr: [3] }).error.issues[1].path).toEqual(['arr', 'every']);
+        expect(schema.safeParse({ arr: [3] }).error.issues[2].path).toEqual(['arr', 'some']);
+        expect(schema.safeParse({ arr: [1] }).error.issues[0].path).toEqual(['arr', 'every']);
+        expect(schema.safeParse({ arr: [4] }).error.issues[0].path).toEqual(['arr']);
+        expect(schema.safeParse({ arr: [4] }).error.issues[1].path).toEqual(['arr', 'every']);
+        expect(schema.safeParse({ arr: [4] }).error.issues[2].path).toEqual(['arr', 'some']);
+        expect(schema.safeParse({ arr: [1, 2, 3] }).success).toBeTruthy();
+    });
+
     it('full-text search', async () => {
         const model = `
         datasource db {
@@ -747,6 +788,11 @@ describe('Zod plugin tests', () => {
             provider = 'prisma-client-js'
         }
     
+        plugin enhancer {
+            provider = '@core/enhancer'
+            output = "$projectRoot/enhance"
+        }
+    
         plugin zod {
             provider = "@core/zod"
             output = "$projectRoot/zod"
@@ -760,7 +806,7 @@ describe('Zod plugin tests', () => {
             password String @omit
         }
         `,
-            { addPrelude: false, pushDb: false, projectDir }
+            { addPrelude: false, pushDb: false, projectDir, getPrismaOnly: true, generateNoCompile: true }
         );
 
         expect(fs.existsSync(path.join(projectDir, 'zod', 'test.txt'))).toBeFalsy();
@@ -781,6 +827,11 @@ describe('Zod plugin tests', () => {
         generator js {
             provider = 'prisma-client-js'
         }
+
+        plugin enhancer {
+            provider = '@core/enhancer'
+            output = "$projectRoot/enhance"
+        }
     
         plugin zod {
             provider = "@core/zod"
@@ -795,7 +846,7 @@ describe('Zod plugin tests', () => {
             password String @omit
         }
         `,
-                { addPrelude: false, pushDb: false, projectDir }
+                { addPrelude: false, pushDb: false, projectDir, generateNoCompile: true }
             )
         ).rejects.toThrow('already exists and is not a directory');
     });