+++ packages/ide/jetbrains/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 ### Added
+- A new `path` parameter to the `@@validate` attribute for providing an optional path to the field that caused the error.
+
+## 2.4.0
+### Added
+- The `uuid()` function is updated to support the new UUID version feature from Prisma.
+
+## 2.3.0 
+### Added
 - New `check()` policy rule function.
 
 ### Fixed